+++ components/flow-orchestration-framework/org.wso2.carbon.identity.flow.execution.engine/src/main/java/org/wso2/carbon/identity/flow/execution/engine/model/ExecutorResponse.java
@@ -27,7 +27,6 @@ import java.util.Map;
 public class ExecutorResponse {
 
     private String result;
-    private String userId;
     private List<String> requiredData;
     private List<String> optionalData;
     private Map<String, Object> updatedUserClaims;
@@ -54,16 +53,6 @@ public class ExecutorResponse {
     public void setResult(String result) {
 
         this.result = result;
-    }
-
-    public String getUserId() {
-
-        return userId;
-    }
-
-    public void setUserId(String userId) {
-
-        this.userId = userId;
     }
 
     public List<String> getRequiredData() {
